requirements.txt
arnparse==0.0.2
boto3
netaddr==0.7.19
dataclasses_json==0.5.2
PyYAML>=5.3.1
packaging
pulumi-policy>=1.3.0,<2.0.0
pulumi>=3.0.0,<4.0.0
pulumi-aws>=4.0.0,<5.0.0
cloudrail-knowledge
